backend/requirements.txt
blinker==1.9.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.1; python_version >= '3.9' and python_version < '4.0'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
lxml==5.3.1; python_version >= '3.6'
markupsafe==3.0.2; python_version >= '3.9'
pymupdf==1.25.4; python_version >= '3.9'
python-docx==1.1.2; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'